Added support for custom actor batch sizes through use of a custom method on actors "fun _batch():USize => 20".  If an actor defines a _batch function, then the actor class will query that function to determine what the batch size for his actor should be.  If no batch function exists, then the default PONY_SCHED_BATCH will be used.
Setting a custom batch size is useful in a number of situations. For
example, if the messages being sent to this actor are very large, then
you might want the system to not allow very many of them to be stored
in the actor's mailbox at one time (by setting a lower batch size, the
actor will become overloaded sooner and mute the sender).  Another
potential use case is in many-to-one architectures, where you might
want single consumer to have a much larger batch size to delay muting
of the many.