In some actor networks, the producer is very quick and the consumer is very slow. In stock pony, actors only being overloaded and mute the producer after the consumer has processed more than one batch worth of messages.  Meanwhile, the producer is free to send untold number of messages and pollute the consumer's mailbox.
With this change, when an actor sends another actor a message they
check if the target actor's mailbox is full.  If it is, the message is
sent but the producing actor mutes itself to prevent any more messages
from polluting the target's mailbox. In addition, the target actor is
immediatetly set to be overloaded.