added support for linking an Info.plist file into built executable
To run an executable as a Mac app, it either needs to be contained in
a bundle or the Info.plist needs to be embedded in the executable file.
This is easily done using a linker flag, so we use the source
translation feature to recognize the plist file and flag the program_t
object to link to the Info.plist.