added concept of opaque and transparent structs
"Opaque" structs are easier to deal with in Pony for FFI, as the
APIs which use them just give pointer/take pointer. Opaque structs
are transpiled to the primitive.

Some APIs require a "transparent" struct. In other words, they require
the Pony program to provide a pointer to an already existing struct.
Hence the transpiler needs to provide a full struct definition for the
struct. Structs are less convenient to work with than opaque primitives.

This code attempts to set rules as to what constitutes the creation of
an opaque struct vs a transparent struct. Right now there are two
criteria a struct must meet in order to be worth the trouble of
transparency.

1. They must be used as a pointer argument to a function
2. They must contain a struct definition with fields