repository now contains compressed+patched llvm 7.1.0 sources
For the llvm-mac build.sh, it used to download the sources for llvm7.1.0
from the release site. Then it would patch the downloaded sources with
the patches used by MacPorts.

Now we include lzip compressed archives of the already compressed sources
directly in the repo. This removes the need for additional downloads and
makes everything a little more streamlined.