first working version. It works with a combination of two things:
1. it makes additions to how TK_AS is handled. If it detects itself
used in specific cases in an if, it inserts a match which returns
true or false.

2. If #1 happens, I walk the AST inside of the if body to find references
to this variable.  I flag these ast nodes with "use this child of the
union type for this reference only", and made changes to reach.c and
lookup.c to make them smart enough to treat this as a narrowed scope.